[Refactor] Move Page Components to Folder
