+++ Frontend/src/app/directives/has-role.directive.ts
