+++ Frontend/src/app/pages/app.routes.ts
@@ -1,17 +1,17 @@
 import { Routes } from '@angular/router';
-import { UserDirectoryComponent } from './components/user-directory/user-directory.component';
-import { MainframeComponent } from './components/mainframe/mainframe.component';
-import { EmptyComponent } from './components/empty/empty.component';
-import { SignInComponent } from './components/signin/signin.component';
-import { RegisterComponent } from './components/register/register.component';
-import { signedInGuard } from './guards/signedIn.guard';
-import { SandboxComponent } from './components/sandbox/sandbox.component';
-import { ServerErrorComponent } from './components/server-error/server-error.component';
-import { NotFoundComponent } from './components/not-found/not-found.component';
-import { LoremIpsumComponent } from './components/lorem-ipsum/lorem-ipsum.component';
-import { UserProfileComponent } from './components/users/user-profile/user-profile.component';
-import { AdminPanelComponent } from './components/admin/admin-panel/admin-panel.component';
-import { adminGuard } from './guards/admin.guard';
+import { UserDirectoryComponent } from '../components/user-directory/user-directory.component';
+import { MainframeComponent } from '../components/mainframe/mainframe.component';
+import { EmptyComponent } from '../components/empty/empty.component';
+import { SignInComponent } from '../components/signin/signin.component';
+import { RegisterComponent } from '../components/register/register.component';
+import { signedInGuard } from '../guards/signedIn.guard';
+import { SandboxComponent } from '../components/sandbox/sandbox.component';
+import { ServerErrorComponent } from '../components/server-error/server-error.component';
+import { NotFoundComponent } from '../components/not-found/not-found.component';
+import { LoremIpsumComponent } from '../components/lorem-ipsum/lorem-ipsum.component';
+import { UserProfileComponent } from '../components/users/user-profile/user-profile.component';
+import { AdminPanelComponent } from '../components/admin/admin-panel/admin-panel.component';
+import { adminGuard } from '../guards/admin.guard';
 
 export const routes: Routes = [
   {
